Refactor Segment events
