+++ resolvers/donationResolver.ts
@@ -5,7 +5,7 @@ import { getTokenPrices, getOurTokenList } from 'monoswap';
 import { Donation } from '../entities/donation';
 import { MyContext } from '../types/MyContext';
 import { Project } from '../entities/project';
-import { getAnalytics } from '../analytics';
+import { getAnalytics, SegmentEvents } from '../analytics';
 import { Token } from '../entities/token';
 import { Repository, In } from 'typeorm';
 import { User } from '../entities/user';
@@ -199,7 +199,7 @@ export class DonationResolver {
         };
 
         analytics.track(
-          'Made donation',
+          SegmentEvents.MADE_DONATION,
           originUser.segmentUserId(),
           segmentDonationMade,
           originUser.segmentUserId(),
@@ -218,7 +218,7 @@ export class DonationResolver {
         };
 
         analytics.track(
-          'Donation received',
+          SegmentEvents.DONATION_RECEIVED,
           projectOwner.segmentUserId(),
           segmentDonationReceived,
           projectOwner.segmentUserId(),
