+++ resolvers/projectResolver.ts
@@ -8,7 +8,7 @@ import {
 } from '../entities/project';
 import { InjectRepository } from 'typeorm-typedi-extensions';
 import { ProjectStatus } from '../entities/projectStatus';
-import { ProjectInput, ImageUpload } from './types/project-input';
+import { ImageUpload, ProjectInput } from './types/project-input';
 import { PubSubEngine } from 'graphql-subscriptions';
 import { pinFile } from '../middleware/pinataUtils';
 import { UserPermissions } from '../permissions';
@@ -17,17 +17,15 @@ import { Donation } from '../entities/donation';
 import { ProjectImage } from '../entities/projectImage';
 import { triggerBuild } from '../netlify/build';
 import { MyContext } from '../types/MyContext';
-import { getAnalytics } from '../analytics';
+import { getAnalytics, SegmentEvents } from '../analytics';
 import { Max, Min } from 'class-validator';
 import { User } from '../entities/user';
 import { Context } from '../context';
 import { Repository } from 'typeorm';
-import { Raw } from 'typeorm';
 import { Service } from 'typedi';
 import config from '../config';
 import slugify from 'slugify';
 import Logger from '../logger';
-import { getProvider, NETWORK_IDS } from '../provider';
 import {
   Arg,
   Args,
@@ -44,18 +42,15 @@ import {
   registerEnumType,
   Resolver,
 } from 'type-graphql';
-
-const analytics = getAnalytics();
-
-import { inspect } from 'util';
 import { errorMessages } from '../utils/errorMessages';
 import {
-  getSimilarTitleInProjectsRegex,
   isWalletAddressSmartContract,
   validateProjectTitle,
   validateProjectTitleForEdit,
   validateProjectWalletAddress,
 } from '../utils/validators/projectValidator';
+
+const analytics = getAnalytics();
 
 @ObjectType()
 class AllProjects {
@@ -617,7 +612,7 @@ export class ProjectResolver {
       description: projectInput?.description?.replace(/<img .*?>/g, ''),
     };
     analytics.track(
-      'Project created',
+      SegmentEvents.PROJECT_CREATED,
       `givethId-${ctx.req.user.userId}`,
       segmentProject,
       null,
@@ -671,7 +666,7 @@ export class ProjectResolver {
     const save = await ProjectUpdate.save(update);
 
     analytics.track(
-      'Project updated - owner',
+      SegmentEvents.PROJECT_UPDATED_OWNER,
       `givethId-${user.userId}`,
       projectUpdateInfo,
       null,
@@ -705,7 +700,7 @@ export class ProjectResolver {
         firstName: donor.firstName,
       };
       analytics.track(
-        'Project updated - donor',
+        SegmentEvents.PROJECT_UPDATED_DONOR,
         `givethId-${donor.id}`,
         donorUpdateInfo,
         null,
@@ -1006,7 +1001,7 @@ export class ProjectResolver {
           };
 
           analytics.track(
-            'Project deactivated',
+            SegmentEvents.PROJECT_DEACTIVATED,
             `givethId-${ctx.req.user.userId}`,
             segmentProject,
             null,
@@ -1059,7 +1054,7 @@ export class ProjectResolver {
       const projectOwner = await User.findOne({ id: Number(project.admin) });
       if (listed === false) {
         analytics.track(
-          'Project unlisted',
+          SegmentEvents.PROJECT_UNLISTED,
           `givethId-${project.admin}`,
           {
             id: project.id,
@@ -1075,7 +1070,7 @@ export class ProjectResolver {
         );
       } else if (listed === true) {
         analytics.track(
-          'Project listed',
+          SegmentEvents.PROJECT_LISTED,
           `givethId-${project.admin}`,
           {
             id: project.id,
